% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-summary-statistics.R
\name{compute_pdf_and_cdf}
\alias{compute_pdf_and_cdf}
\title{Density and Cumulative Distribution Function}
\usage{
compute_pdf_and_cdf(margpost, finegrid = NULL)
}
\arguments{
\item{margpost}{The output of \code{aghq::marginal_posterior}. See the documentation for that function.}

\item{finegrid}{Optional, a grid of values on which to compute the CDF. The default makes
use of the values in \code{margpost} but if the results are unsuitable, you may wish to
modify this manually.}
}
\value{
A tbl_df/tbl/data.frame with columns \code{theta}, \code{pdf} and \code{cdf} corresponding
to the value of the parameter and its estimated PDF and CDF at that value.
}
\description{
Compute the density and cumulative distribution function of the approximate posterior.
The density is approximated on a find grid using a polynomial interpolant.
The CDF can't be computed exactly (if it could, you wouldn't be using quadrature!),
so a fine grid is laid down and the CDF is approximated at each grid point
using a simpler integration rule and a polynomial interpolant. This method tends
to work well, but won't always.
}
\seealso{
Other summaries: 
\code{\link{compute_moment}()},
\code{\link{compute_quantiles}()},
\code{\link{interpolate_marginal_posterior}()},
\code{\link{marginal_posterior}()}
}
\concept{summaries}
