% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-summary-statistics.R
\name{compute_quantiles}
\alias{compute_quantiles}
\title{Quantiles}
\usage{
compute_quantiles(margpost, q = c(0.025, 0.975))
}
\arguments{
\item{margpost}{The output of \code{aghq::marginal_posterior}. See the documentation for that function.}

\item{q}{Numeric vector of values in (0,1). The quantiles to compute.}
}
\value{
A named numeric vector containing the quantiles you asked for, for the
variable whose marginal posterior you provided.
}
\description{
Compute marginal quantiles using AGHQ. This function works by first approximating
the CDF using \code{aghq::compute_pdf_and_cdf} and then inverting the approximation numerically.
}
\seealso{
Other summaries: 
\code{\link{compute_moment}()},
\code{\link{compute_pdf_and_cdf}()},
\code{\link{interpolate_marginal_posterior}()},
\code{\link{marginal_posterior}()}
}
\concept{summaries}
