% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-summary-statistics.R
\name{default_transformation}
\alias{default_transformation}
\title{Default transformation}
\usage{
default_transformation()
}
\description{
Default (identity) transformation object. Default argument in package functions
which accept transformations, and useful for user inspection.
}
\examples{

default_transformation()

}
\seealso{
Other transformations: 
\code{\link{make_transformation}()},
\code{\link{validate_transformation}()}
}
\concept{transformations}
