% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcdatalist}
\alias{gcdatalist}
\title{Transformed Globular Clusters data}
\format{
An object of class \code{list} of length 6.
}
\source{
Eadie GM, Harris WE (2016). “Bayesian mass estimates of the Milky Way: the dark and light sides of parameter assumptions.” The Astrophysical Journal, 829(108).
}
\usage{
gcdatalist
}
\description{
GC data prepared for input into the TMB template, for purposes of example.
There are a lot of example-specific data preprocessing steps that are not
related to the AGHQ method, so for brevity these are done beforehand.
}
\keyword{datasets}
