% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-summary-statistics.R
\name{interpolate_marginal_posterior}
\alias{interpolate_marginal_posterior}
\title{Interpolate the Marginal Posterior}
\usage{
interpolate_marginal_posterior(margpost)
}
\arguments{
\item{margpost}{The output of \code{aghq::marginal_posterior}. See the documentation for that function.}
}
\value{
A function of \code{theta} which computes the log interpolated normalized marginal posterior.
}
\description{
Build a Lagrange polynomial interpolant of the marginal posterior, for plotting
and for computing quantiles.
}
\seealso{
Other summaries: 
\code{\link{compute_moment}()},
\code{\link{compute_pdf_and_cdf}()},
\code{\link{compute_quantiles}()},
\code{\link{marginal_posterior}()}
}
\concept{summaries}
